backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
pyjwt==2.8.0
redis==5.0.4
sqlalchemy==2.0.30
aiofiles==23.2.1
python-dotenv==1.0.1
transitions==0.9.1
PyYAML==6.0.1
pytest==8.2.0
pytest-asyncio==0.23.7
aiosqlite==0.20.0
sse-starlette==2.1.0
